+++ dotCMS/src/main/java/com/dotcms/rest/api/v1/content/dotimport/ContentImportHelper.java
@@ -1,4 +1,4 @@
-package com.dotcms.rest.api.v1.content._import;
+package com.dotcms.rest.api.v1.content.dotimport;
 
 import com.dotcms.jobs.business.api.JobQueueManagerAPI;
 import com.dotcms.rest.api.v1.JobQueueManagerHelper;
