+++ dotcms-integration/src/test/java/com/dotcms/Junit5Suite1.java
@@ -4,7 +4,7 @@ import com.dotcms.jobs.business.api.JobQueueManagerAPICDITest;
 import com.dotcms.jobs.business.api.JobQueueManagerAPIIntegrationTest;
 import com.dotcms.jobs.business.processor.impl.ImportContentletsProcessorIntegrationTest;
 import com.dotcms.jobs.business.queue.PostgresJobQueueIntegrationTest;
-import com.dotcms.rest.api.v1.content._import.ContentImportResourceIntegrationTest;
+import com.dotcms.rest.api.v1.content.dotimport.ContentImportResourceIntegrationTest;
 import com.dotcms.rest.api.v1.job.JobQueueHelperIntegrationTest;
 import org.junit.platform.suite.api.SelectClasses;
 import org.junit.platform.suite.api.Suite;
